StunClient Added
Addition of a StunClient working in UDP (most common) or TCP (ejabberd)
Messages and Attributes are automically converted from byte array to
objects through the use of implicit operators
Constants are big-endian defined in the code , this was made to match to
the RFC5389 where these constants are written in big-endian (network
byte-order)
The drawbak is that Windows uses little-endian encoding and that's why
there are methods used to convert these constants.
Bytes that are never reversed are the Attributes values and the
TransactionID
This code is already working with ejabberd 2.1.10 (TCP and UDP) and some
free Stun servers available on the Internet (UDP only).
StunClient contains a sample code in its Main static method.